Refactor OptiGlue 1.18
