+++ OptiGlue/1.18/src/main/resources/fabric.mod.json
@@ -18,7 +18,7 @@
     "optigui": [
       {
         "adapter": "kotlin",
-        "value": "opekope2.optiglue.OptiGlueMod"
+        "value": "opekope2.optiglue_1_18.OptiGlueMod"
       }
     ]
   },
